Add files via upload
We start with the basics of ERP SYSTEMS.

Business Functions and Processes
Enterprise Resource Planning (ERP) systems are core software programs used by companies to
integrate and coordinate information in every area of the business.
 business process is a collection of activities that takes
one or more kinds of input and creates an output, such as a report or forecast, that is of value to the
customer. ERP software supports the efficient operation of business processes by integrating tasks
related to sales, marketing, manufacturing, logistics, accounting, and staffing—throughout a
business.FUNCTIONAL AREAS AND BUSINESS PROCESSES
To understand ERP, you must first understand how a business works. Let’s begin by
looking at a typical business’s areas of operation. These areas, called functional areas of
operation, are broad categories of business activities.
Functional Areas of Operation
Most companies have four main functional areas of operation: Marketing and Sales (M/S),
Supply Chain Management (SCM), Accounting and Finance (A/F), and Human Resources
(HR). Each area is composed of a variety of narrower business functions, which are
activities specific to that functional area of operation. 

Historically, businesses have had organizational structures that separated the
functional areas. Business schools have been similarly organized, so each functional area
has been taught as a separate course. In a company separating functional areas in this
way, Marketing and Sales might be completely isolated from Supply Chain Management,
even though the Marketing and Sales staff sell what the employees in Supply Chain
Management procure and produce. Thus, you might conclude that what happens in one
functional area is not closely related to what happens in others. As you will learn in this
chapter, however, functional areas are interdependent, each requiring data from the
others. The better a company can integrate the activities of each functional area, the
more successful it will be in today’s highly competitive environment. The Association to
Advance Collegiate Schools of Business (AACSB)—the accreditation board of university
business schools—is now promoting integration between functional areas of business for
higher education.
Integration also contributes to improvements in communication and workflow. Each
area’s information system depends on data from other functional areas. An information
system (IS) includes the people, procedures, software, and computers that store, organize,
analyze, and deliver information. This chapter illustrates the need for information sharing
between functional areas and the effects on the business if this information is not
integrated. In later sections, you will also see some examples of typical business processes
and how these processes routinely cross functional areas.

Business Processes
More managers are now thinking in terms of business processes rather than business
functions. Recall that a business process is a collection of activities that takes one or more
kinds of input and creates an output that is of value to the customer. The customer for a
business process may be a traditional external customer (the person who buys the finished
product), or it may be an internal customer (such as a colleague in another department).
For example, what is sold through Marketing and Sales is linked to what is procured and
produced by Supply Chain Management.